@@ -11,11 +11,11 @@
         }
     ],
     "require": {
-        "php": "^7.3",
+        "php": "^7.4",
         "ext-json": "*",
         "keboola/csv": "^2.2",
         "keboola/db-adapter-snowflake": "1.0.4",
-        "keboola/db-extractor-common": "^13.3",
+        "keboola/db-extractor-common": "^14.2",
         "keboola/db-extractor-config": "^1.4",
         "keboola/db-extractor-table-format": "^3.1",
         "keboola/php-component": "^8.1",
